Remove unused code replaced with new implemenations for translations
