+++ application/controllers/handler.js
@@ -18,9 +18,6 @@ const boom = require('boom'),
     Joi = require('joi'),
     async = require('async');
 
-// TODO remove this from here after we've refactored all database-specific code into slide/deck database js files
-const ChangeLog = require('../lib/ChangeLog');
-
 const tagService = require('../services/tag');
 const fileService = require('../services/file');
 
@@ -106,70 +103,6 @@ let self = module.exports = {
             request.log('error', err);
             reply(boom.badImplementation());
         });
-
-    },
-
-    //updates slide by creating a new revision
-    // DEPRECATED
-    updateSlide: function(request, reply) {
-        let userId = request.auth.credentials.userid;
-        let slideId = request.params.id;
-
-        // fill in the user id from auth
-        request.payload.user = userId;
-
-        { // these brackets are kept during handleChange removal to keep git blame under control
-
-            deckDB.getActiveRevisionFromDB(request.payload.root_deck).then((parentDeckId) => {
-
-                //update root deck with active revision
-                if(parentDeckId)
-                    request.payload.root_deck = parentDeckId;
-
-                //create the slide revision in the database
-                return slideDB.replace(slideId, request.payload).then((slideRef) => {
-                    // send tags to tag-service
-                    if(request.payload.tags && request.payload.tags.length > 0){
-                        tagService.upload(request.payload.tags, userId).catch( (e) => {
-                            request.log('warning', 'Could not save tags to tag-service for slide ' + slideId + ': ' + e.message);
-                        });
-                    }
-
-                    //we must update all decks in the 'usage' attribute
-                    return slideDB.get(slideRef.id).then((newSlide) => {
-
-                        // prepare the newSlide response object
-                        newSlide.revisions = [newSlide.revisions[newSlide.revisions.length-1]];
-
-                        // update the content item of the parent deck with the new revision id
-                        return deckDB.updateContentItem(newSlide, '', request.payload.root_deck, 'slide', userId, request.payload.top_root_deck)
-                        .then(({updatedDeckRevision}) => {
-                            // the updateContentItem returns, amongs other things, the updated revision of the parent (root_deck)
-                            // we need this to have access to the theme for the new updated slide
-
-                            //create thumbnail for the new slide revision
-                            let content = newSlide.revisions[0].content, newSlideId = newSlide._id+'-'+newSlide.revisions[0].id;
-                            if(content === ''){
-                                content = '<h2>'+newSlide.revisions[0].title+'</h2>';
-                                //for now we use hardcoded template for new slides
-                                content = slidetemplate;
-                            }
-                            fileService.createThumbnail(content, newSlideId, updatedDeckRevision.theme).catch((err) => {
-                                request.log('warn', `could not create thumbnail for updated slide ${newSlideId}: ${err.message || err}`);
-                            });
-
-                            reply(newSlide);
-                        });
-
-                    });
-                });
-
-            }).catch((error) => {
-                request.log('error', error);
-                reply(boom.badImplementation());
-            });
-
-        }
 
     },
 
@@ -756,699 +689,6 @@ let self = module.exports = {
                 reply(boom.badImplementation());
             });
         }
-    },
-
-    // authorize node creation and iterate nodeSpec array to apply each insert
-    // DEPRECATED
-    createDeckTreeNodeWithCheck: function(request, reply) {
-        let userId = request.auth.credentials.userid;
-        let rootDeckId = request.payload.selector.id;
-
-        // TODO proper authorization checking the actual parent id
-        return authorizeUser(userId, rootDeckId, rootDeckId).then((boomError) => {
-            if (boomError) return reply(boomError);
-
-            let nodeSpecs = request.payload.nodeSpec;
-            if (nodeSpecs.length < 2) {
-                request.payload.nodeSpec = nodeSpecs[0];
-                return self.createDeckTreeNode(request, reply);
-            }
-
-            // do some complex validations
-            // we only support more than one nodeSpec when attaching,
-            // so isMove cannot be true
-            if (request.payload.isMove) {
-                return reply(boom.badData());
-            }
-
-            // also, all ids should be valid numbers
-            if (!nodeSpecs.every((node) => node.id && node.id !== '0')) {
-                return reply(boom.badData());
-            }
-
-            // check if we append at the end (no position argument) or at a position
-            let reverseOrder = (request.payload.selector.stype === 'slide');
-            if (reverseOrder) {
-                // if we *don't* attach to the end, we need to
-                // reverse the node specs because they are added right after
-                // the position specified in selector, like in a stack (LIFO)
-                // we would like to provide the semantics of a queue (FIFO)
-                nodeSpecs.reverse();
-            } else {
-                // if appending to deck lets remove the spath because we always attach to the last position
-                request.payload.selector.spath = '';
-            }
-
-            async.concatSeries(nodeSpecs, (nodeSpec, done) => {
-                // just put this nodespec
-                request.payload.nodeSpec = nodeSpec;
-
-                self.createDeckTreeNode(request, (result) => {
-                    // an error already logged
-                    if (result && result.isBoom) done(result);
-
-                    // result is not an error
-                    done(null, result);
-                });
-            }, (err, results) => {
-                if (err) {
-                    // an error already logged
-                    if (err.isBoom) {
-                        reply(err);
-                    } else {
-                        // an error in this method code, not logged
-                        request.log('error', err);
-                        reply(boom.badImplementation());
-                    }
-                } else {
-                    // if needed, we again reverse the results to match the node spec order
-                    if (reverseOrder) results.reverse();
-
-                    reply(results);
-                }
-            });
-        });
-
-    },
-
-    //creates a node (deck or slide) into the given deck tree
-    // DEPRECATED
-    createDeckTreeNode: function(request, reply) {
-        let node = {};
-        let top_root_deck = request.payload.selector.id;
-        let userId = request.auth.credentials.userid;
-
-        //check if it is a slide or a deck
-        if(request.payload.nodeSpec.type === 'slide'){
-            if(request.payload.nodeSpec.id && request.payload.nodeSpec.id !== '0'){
-                //it means it is an existing node, we should retrieve the details
-                let spath = request.payload.selector.spath;
-                let spathArray = spath.split(';');
-                let parentID, slidePosition;
-                if(spathArray.length > 1){
-
-                    let parentArrayPath = spathArray[spathArray.length-2].split(':');
-                    parentID = parentArrayPath[0];
-
-                } else if (request.payload.selector.stype === 'deck') {
-                    parentID = request.payload.selector.sid;
-                } else {
-                    // means we are at root deck
-                    parentID = request.payload.selector.id;
-                }
-
-                let slideArrayPath = spathArray[spathArray.length-1].split(':');
-                slidePosition = parseInt(slideArrayPath[1])+1;
-
-                // TODO remove this deprecated part of the code
-                if (request.payload.isMove) {
-                    let slideRevision = parseInt(request.payload.nodeSpec.id.split('-')[1])-1;
-                    self.getSlide({
-                        'params' : {'id' : request.payload.nodeSpec.id.split('-')[0]},
-                        'log': request.log.bind(request),
-                    }, (slide) => {
-                        if (slide.isBoom) return reply(slide);
-
-                        //change position of the existing slide
-                        slide.id = slide._id;
-
-                        deckDB.insertNewContentItem(slide, slidePosition, parentID, 'slide', slideRevision+1, userId, top_root_deck)
-                        .then((updatedDeckRevision) => {
-                            // since we moved the slide maybe it's under a deck with a different theme, so let's create the thumbnail as well
-                            let slideId = slide.id+'-'+(slideRevision+1);
-                            fileService.createThumbnail(slide.revisions[slideRevision].content, slideId, updatedDeckRevision.theme).catch((err) => {
-                                request.log('warn', `could not create thumbnail for slide duplicate ${slideId}: ${err.message || err}`);
-                            });
-
-                            return slideDB.addToUsage({ref:{id:slide._id, revision: slideRevision+1}, kind: 'slide'}, parentID.split('-')).then(() => {
-                                node = {title: slide.revisions[slideRevision].title, id: slide.id+'-'+slide.revisions[slideRevision].id, type: 'slide'};
-                                reply(node);
-                            });
-                        }).catch( (err) => {
-                            request.log('error', err);
-                            reply(boom.badImplementation());
-                        });
-
-                    });
-
-                    // finished
-                    return;
-                }
-
-                // we are attaching a copy of a slide that may or may not be in the current tree
-                let slideId = request.payload.nodeSpec.id;
-                // let's keep the exact action tracked
-                let addAction = 'attach';
-                if (request.payload.nodeSpec.id === request.payload.selector.sid) {
-                    // this means we create a slide duplicate and insert after the slide in selector
-                    addAction = 'copy';
-                }
-
-                // if nodeSpec.root is defined, means it's an external attach of a slide in a deck
-                let slideRootId = request.payload.nodeSpec.root;
-                if (slideRootId) {
-                    // we attach the slide with its variants (translations)
-                    slideDB.findSlideNode(slideRootId, slideId).then((slideNode) => {
-                        if (!slideNode) {
-                            return reply(boom.badData(`could not find slide: ${slideId} in deck tree: ${slideRootId}`));
-                        }
-
-                        // we must duplicate the slide node
-                        return slideDB.copySlideNode(slideRootId, slideId, parentID, userId).then((newContentItem) => {
-                            return deckDB.insertContentItem(newContentItem, slidePosition, parentID, userId, top_root_deck, addAction).then((updatedDeckRevision) => {
-                                let theme = updatedDeckRevision.theme;
-                                return slideDB.getContentItemSlides(newContentItem).then((slides) => {
-                                    // generate thumbnails but don't wait for it
-                                    for (let slide of slides) {
-                                        let newSlideId = util.toIdentifier(slide);
-                                        fileService.createThumbnail(slide.content, newSlideId, theme).catch((err) => {
-                                            console.warn(`could not create thumbnail for slide ${newSlideId}, error was: ${err.message}`);
-                                        });
-                                    }
-
-                                    // the node data are the same as the original
-                                    reply({ title: slideNode.slide.title, id: util.toIdentifier(newContentItem.ref), type: 'slide' });
-                                });
-
-                            });
-
-                        });
-
-                    }).catch((err) => {
-                        request.log('error', err);
-                        reply(boom.badImplementation());
-                    });
-
-                } else {
-                    // otherwise, we keep things backwards-compatible and insert a simple slide copy.
-                    // in this case the slide copy will have the same language as the deck tree we are attaching to
-                    slideDB.getSlideRevision(slideId).then((slide) => {
-                        if (!slide) {
-                            return reply(boom.badData(`could not locate slide to attach: ${slideId}`));
-                        }
-
-                        return slideDB.copy(slide, parentID, userId).then((inserted) => {
-                            // it's slide copy, so revision is 1
-                            let newContentItem = { ref: { id: inserted._id, revision: 1 }, kind: 'slide' };
-                            return deckDB.insertContentItem(newContentItem, slidePosition, parentID, userId, top_root_deck, addAction).then((updatedDeckRevision) => {
-                                // we can now pick the theme of the parent deck and create the thumbnail!
-                                let newSlideId = util.toIdentifier(newContentItem.ref);
-                                fileService.createThumbnail(inserted.revisions[0].content, newSlideId, updatedDeckRevision.theme).catch((err) => {
-                                    console.warn(`could not create thumbnail for slide ${newSlideId}, error was: ${err.message}`);
-                                });
-
-                                // the node data are the same as the original
-                                reply({ title: slide.title, id: newSlideId, type: 'slide' });
-                            });
-                        });
-
-                    }).catch((err) => {
-                        request.log('error', err);
-                        reply(boom.badImplementation());
-                    });
-
-                }
-
-            }else{
-                //need to make a new slide
-                let spath = request.payload.selector.spath;
-                let spathArray = spath.split(';');
-                let parentID, slidePosition;
-                if(spathArray.length > 1){
-
-                    let parentArrayPath = spathArray[spathArray.length-2].split(':');
-                    parentID = parentArrayPath[0];
-
-                }
-                else{
-                    parentID = request.payload.selector.id;
-                }
-                let slideArrayPath = spathArray[spathArray.length-1].split(':');
-                slidePosition = parseInt(slideArrayPath[1])+1;
-                if(request.payload.selector.stype === 'deck'){
-                    //selector is deck, we can get the root deck id directly
-                    parentID = request.payload.selector.sid;
-                    slidePosition = 0;
-                }
-
-                { // these brackets are kept during handleChange removal to keep git blame under control
-
-                    self.getDeck({
-                        'params': {'id':parentID},
-                        'log': request.log.bind(request),
-                    }, (parentDeck) => {
-                        if (parentDeck.isBoom) return reply(parentDeck);
-
-                        let slide = {
-                            'title': 'New slide',
-                            'content': slidetemplate,
-                            'allowMarkdown': false,
-                            'markdown': '',
-                            'language': parentDeck.revisions[0].language,
-                            'license': parentDeck.license,
-                            'root_deck': parentID,
-                            'position' : slidePosition,
-                            'user': userId,
-                            'theme' : request.payload.theme,
-                        };
-
-                        if (parentDeck.slideDimensions) {
-                            slide.dimensions = parentDeck.slideDimensions;
-                        }
-                        if(request.payload.hasOwnProperty('content')){
-                            slide.content = request.payload.content;
-                        }
-                        if(request.payload.hasOwnProperty('markdown')){
-                            slide.markdown = request.payload.markdown;
-                        }
-                        if(request.payload.hasOwnProperty('allowMarkdown')){
-                            slide.allowMarkdown = request.payload.allowMarkdown;
-                        }
-                        if(request.payload.hasOwnProperty('title')){
-                            slide.title = request.payload.title;
-                        }
-                        if(request.payload.hasOwnProperty('license')){
-                            slide.license = request.payload.license;
-                        }
-                        if(request.payload.hasOwnProperty('speakernotes')){
-                            slide.speakernotes = request.payload.speakernotes;
-                        }
-
-                        // create the new slide into the database
-
-                        // insert the slide
-                        slideDB.insert(slide).then((inserted) => {
-                            // empty results means something wrong with the payload
-                            if (!inserted) throw boom.badData();
-
-                            let createdSlide = co.rewriteID(inserted);
-
-                            node = {title: createdSlide.revisions[0].title, id: createdSlide.id+'-'+createdSlide.revisions[0].id, type: 'slide'};
-
-                            //we have to return from the callback, else empty node is returned because it is updated asynchronously
-                            return deckDB.insertNewContentItem(createdSlide, slidePosition, parentID, 'slide', 1, userId, top_root_deck).then((updatedDeckRevision) => {
-
-                                // slide is now inserted, so we can create the thumbnail using the (direct) parent deck theme
-
-                                // create thumbnail from the newly created slide revision
-                                let content = createdSlide.revisions[0].content, slideId = createdSlide.id+'-'+1;
-                                if (content === '') {
-                                    // content = '<h2>'+inserted.revisions[0].title+'</h2>';
-                                    // TODO for now we use hardcoded template for new slides
-                                    content = slidetemplate;
-                                }
-
-                                // don't wait for it before returning
-                                fileService.createThumbnail(content, slideId, updatedDeckRevision.theme).catch((err) => {
-                                    request.log('warn', `could not create thumbnail for new slide ${slideId}: ${err.message || err}`);
-                                });
-
-                                reply(node);
-                            });
-
-                        }).catch((err) => {
-                            if (err.isBoom) return reply(err);
-
-                            request.log('error', err);
-                            reply(boom.badImplementation());
-                        });
-                    });
-                }
-            }
-        }else{
-            //create a deck node
-            if(request.payload.nodeSpec.id && request.payload.nodeSpec.id !== '0'){
-                //id is specified, it means it is an existing node
-                let spath = request.payload.selector.spath;
-                let spathArray = spath.split(';');
-                let parentID, deckPosition;
-                if(spathArray.length > 1){
-
-                    let parentArrayPath = spathArray[spathArray.length-2].split(':');
-                    parentID = parentArrayPath[0];
-
-                }
-                else{
-                    parentID = request.payload.selector.sid;
-                }
-                if(request.payload.selector.stype === 'slide'){
-                    parentID = request.payload.selector.id;
-                }
-
-                let deckArrayPath = spathArray[spathArray.length-1].split(':');
-                deckPosition = parseInt(deckArrayPath[1])+1;
-                let deckRevision = parseInt(request.payload.nodeSpec.id.split('-')[1])-1;
-
-                //NOTE must check if it is a move action, if not, we are appending an external subdeck and we need to fork it
-                if(request.payload.isMove){
-                    // TODO remove this deprecated part of the code
-                    self.getDeck({
-                        'params': {'id' : request.payload.nodeSpec.id},
-                        'log': request.log.bind(request),
-                    }, (deck) => {
-                        if (deck.isBoom) return reply(deck);
-
-                        deck.id = deck._id;
-
-                        { // these brackets are kept during handleChange removal to keep git blame under control
-
-                            if(request.payload.selector.stype === 'deck'){
-                                parentID = request.payload.selector.sid;
-                            }
-                            else{
-                                parentID = request.payload.selector.id;
-                            }
-
-                            let insertContentItemPromise = deckDB.insertNewContentItem(deck, deckPosition, parentID, 'deck', deckRevision+1, userId, top_root_deck);
-                            let addToUsagePromise = deckDB.addToUsage({ref:{id:deck._id, revision: deckRevision+1}, kind: 'deck'}, parentID.split('-'));
-
-                            Promise.all([insertContentItemPromise, addToUsagePromise]).then( () => {
-                                //we have to return from the callback, else empty node is returned because it is updated asynchronously
-                                self.getDeckTree({
-                                    'params': {'id' : deck.id},
-                                    'log': request.log.bind(request),
-                                }, (deckTree) => {
-                                    if (deckTree.isBoom) return reply(deckTree);
-
-                                    reply(deckTree);
-                                });
-                            }).catch( (err) => {
-                                request.log('error', err);
-                                reply(boom.badImplementation());
-                            });
-
-                        }
-                    });
-                }
-                else{
-                    // find the parent, again
-                    if(request.payload.selector.stype === 'deck'){
-                        parentID = request.payload.selector.sid;
-                    }
-                    else{
-                        parentID = request.payload.selector.id;
-                    }
-
-                    let sourceId = request.payload.nodeSpec.id;
-                    treeDB.attachDeckTree(sourceId, parentID, deckPosition, top_root_deck, userId).then((forkResult) => {
-                        if (!forkResult) {
-                            // source id not found
-                            return reply(boom.badData(`could not locate specified deck: ${sourceId}`));
-                        }
-                        //we have to return from the callback, else empty node is returned because it is updated asynchronously
-                        return treeDB.getDeckTree(forkResult.root_deck).then((deckTree) => {
-                            reply(deckTree);
-                        });
-                    }).catch((err) => {
-                        request.log('error', err);
-                        reply(boom.badImplementation());
-                    });
-                }
-            }else{
-                //id is not specified, we need to make a new deck
-                let spath = request.payload.selector.spath;
-                let spathArray = spath.split(';');
-                let parentID, deckPosition;
-                if(spathArray.length > 1){
-
-                    let parentArrayPath = spathArray[spathArray.length-2].split(':');
-                    parentID = parentArrayPath[0];
-
-                }
-                else{
-                    parentID = request.payload.selector.id;
-                }
-                if(request.payload.selector.stype === 'deck'){
-                    parentID = request.payload.selector.sid;
-                }
-
-                let deckArrayPath = spathArray[spathArray.length-1].split(':');
-                deckPosition = parseInt(deckArrayPath[1])+1;
-
-                let payload = {
-                    title: 'New deck',
-                    description: '',
-                };
-                treeDB.createSubdeck(payload, parentID, deckPosition, top_root_deck, userId).then((newDeck) => {
-                    // this creates an empty subdeck, let's also add a sample slide
-                    // init the payload from the optional first_slide object data
-                    let payload = Object.assign({
-                        title: 'New slide',
-                        content: slidetemplate,
-                        markdown: '',
-                        speakernotes: '',
-                    }, request.payload.first_slide);
-
-                    return treeDB.createSlide(payload, util.toIdentifier(newDeck.ref), 0, top_root_deck, userId).then((newSlide) => {
-                        // we have to return from the callback, else empty node is returned because it is updated asynchronously
-                        return treeDB.getDeckTree(newDeck.ref.id).then((newTree) => {
-                            // let's create a thumbnail but dont wait for it
-                            let slideId = util.toIdentifier(newSlide.ref);
-                            fileService.createThumbnail(payload.content, slideId, newTree.theme).catch((err) => {
-                                request.log('warn', `could not create thumbnail for new slide ${slideId}: ${err.message || err}`);
-                            });
-
-                            reply(newTree);
-                        });
-                    });
-
-                }).catch((err) => {
-                    request.log('error', err);
-                    reply(boom.badImplementation());
-                });
-            }
-        }
-    },
-
-    //renames a decktree node (slide or deck)
-    // DEPRECATED
-    renameDeckTreeNode: function(request, reply) {
-        let userId = request.auth.credentials.userid;
-
-        //check if it is deck or slide
-        if(request.payload.selector.stype === 'deck'){
-            let root_deck = request.payload.selector.sid;
-
-            { // these brackets are kept during handleChange removal to keep git blame under control
-
-                let top_root_deck = request.payload.selector.id;
-                deckDB.rename(root_deck, request.payload.name, {}, top_root_deck, userId).then((renamed) => {
-                    let response = {'title' : renamed};
-                    reply(response);
-                }).catch((error) => {
-                    request.log('error', error);
-                    reply(boom.badImplementation());
-                });
-            }
-
-        }else {
-            //it is a slide, must find root deck id
-            let root_deck ;
-            let slide_id = request.payload.selector.sid;
-            let spath = request.payload.selector.spath;
-            let spathArray = spath.split(';');
-            if(spathArray.length > 1){
-                let parentArrayPath = spathArray[spathArray.length-2].split(':');
-                root_deck = parentArrayPath[0];
-            }
-            else{
-                root_deck = request.payload.selector.id;
-            }
-            //we must create a new slide revision as well, because of renaming it
-            self.getSlide({
-                'params' : {'id' : slide_id},
-                'log': request.log.bind(request),
-            }, (slide) => {
-                if (slide.isBoom) return reply(slide);
-
-                let new_slide = {
-                    'title' : request.payload.name,
-                    'content' : slide.revisions[0].content,
-                    'markdown' : slide.revisions[0].markdown,
-                    'speakernotes' : slide.revisions[0].speakernotes,
-                    'user' : String(userId),
-                    'root_deck' : root_deck,
-                    'top_root_deck' : request.payload.selector.id,
-                    'language' : slide.language,
-                    'license' : slide.license,
-                    'tags' : slide.revisions[0].tags,
-                    'dataSources' : slide.revisions[0].dataSources
-                };
-                if (slide.revisions[0].dimensions) {
-                    new_slide.dimensions = slide.revisions[0].dimensions;
-                }
-                if(new_slide.speakernotes === null){
-                    new_slide.speakernotes = '';
-                }
-                if(new_slide.tags === null){
-                    new_slide.tags = [];
-                }
-                if(new_slide.dataSources === null){
-                    new_slide.dataSources = [];
-                }
-                let forwardedRequest = {
-                    'params' : {'id' :encodeURIComponent(slide_id)},
-                    'payload' : new_slide,
-                    'auth': request.auth,
-                    'log': request.log.bind(request),
-                };
-                self.updateSlide(forwardedRequest, (updated) => {
-                    reply(updated);
-                });
-            });
-        }
-    },
-    //deletes a decktree node by removing its reference from its parent deck (does not actually delete it from the database)
-    // DEPRECATED
-    deleteDeckTreeNode: function(request, reply) {
-        let userId = request.auth.credentials.userid;
-
-        //NOTE no removal in the DB, just unlink from content items, and update the positions of the other elements
-        let spath = request.payload.selector.spath;
-        let spathArray = spath.split(';');
-        let parentID, itemPosition;
-        if(spathArray.length > 1){
-
-            let parentArrayPath = spathArray[spathArray.length-2].split(':');
-            parentID = parentArrayPath[0];
-
-        }
-        else{
-            parentID = request.payload.selector.id;
-        }
-
-        let itemArrayPath = spathArray[spathArray.length-1].split(':');
-        itemPosition = itemArrayPath[1];
-
-        { // these brackets are kept during handleChange removal to keep git blame under control
-
-            //remove link of content item from db
-            let top_root_deck = request.payload.selector.id;
-            deckDB.removeContentItem(itemPosition, parentID, top_root_deck, userId)
-            .then((removed) => {
-                if(!removed){
-                    removed = {};
-                }
-                reply(removed);
-            })
-            .catch((err) => {
-                request.log('error', err);
-                reply(boom.badImplementation());
-            });
-        }
-
-    },
-    //changes position of a deck tree node inside the decktree
-    // DEPRECATED
-    moveDeckTreeNode: function(request, reply) {
-        let userId = request.auth.credentials.userid;
-
-        //first delete the node from its current position
-        self.deleteDeckTreeNode({
-            'payload': {'selector' : request.payload.sourceSelector, 'user': String(userId), 'isMove' : true},
-            'auth': request.auth,
-            'log': request.log.bind(request),
-        },
-        (removed) => {
-            if (removed.isBoom) return reply(removed);
-            //must update revision ids
-            let nodeSpec = {'id': request.payload.sourceSelector.sid, 'type': request.payload.sourceSelector.stype};
-            let sourceParentDeck = request.payload.sourceSelector.id;
-            let spathArray = request.payload.sourceSelector.spath.split(';');
-            if(spathArray.length > 1){
-
-                let parentArrayPath = spathArray[spathArray.length-2].split(':');
-                sourceParentDeck = parentArrayPath[0];
-            }
-            let targetParentDeck = request.payload.targetSelector.id;
-            if(request.payload.targetSelector.spath !== ''){
-                if(request.payload.targetSelector.stype === 'deck'){
-                    targetParentDeck = request.payload.targetSelector.sid;
-                }
-                else{
-                    let targetspathArray = request.payload.targetSelector.spath.split(';');
-                    if(targetspathArray.length > 1){
-
-                        let parentArrayPath = targetspathArray[targetspathArray.length-2].split(':');
-                        targetParentDeck = parentArrayPath[0];
-                    }
-                    else{
-                        let parentArrayPath = targetspathArray[targetspathArray.length-1].split(':');
-                        targetParentDeck = parentArrayPath[0];
-                    }
-                }
-
-            }
-
-            let removed_changeset, inserted_changeset ;
-            if(removed.hasOwnProperty('changeset')){
-                //handle changes coming from the recursive revisioning
-                removed_changeset = removed.changeset;
-                if(removed_changeset.hasOwnProperty('new_revisions')){
-                    for(let i = 0; i < removed_changeset.new_revisions.length; i++){
-                        let next_new_revision = removed_changeset.new_revisions[i];
-                        if(i === 0 && removed_changeset.new_revisions[i].hasOwnProperty('root_changed')){
-                            next_new_revision = removed_changeset.new_revisions[i].root_changed;
-                        }
-                        let next_new_revision_path = next_new_revision.split('-');
-                        if(sourceParentDeck.split('-')[0] === next_new_revision_path[0]){
-                            sourceParentDeck = sourceParentDeck.split('-')[0] + '-' + next_new_revision_path[1];
-                        }
-                        if(targetParentDeck.split('-')[0] === next_new_revision_path[0]){
-                            targetParentDeck = targetParentDeck.split('-')[0] + '-' + next_new_revision_path[1];
-                        }
-                        if(request.payload.targetSelector.sid.split('-')[0] === next_new_revision_path[0]){
-                            request.payload.targetSelector.sid = request.payload.targetSelector.sid.split('-')[0] + '-' + next_new_revision_path[1];
-                        }
-                        if(request.payload.targetSelector.id.split('-')[0] === next_new_revision_path[0]){
-                            request.payload.targetSelector.id = request.payload.targetSelector.id.split('-')[0] + '-' + next_new_revision_path[1];
-                        }
-                        if(nodeSpec.id.split('-')[0] === next_new_revision_path[0]){
-                            nodeSpec.id = nodeSpec.id.split('-')[0] + '-' + next_new_revision_path[1];
-                        }
-                    }
-                }
-            }
-
-            let itemArrayPath = spathArray[spathArray.length-1].split(':');
-            let itemPosition = itemArrayPath[1];
-            if(sourceParentDeck === targetParentDeck && parseInt(itemPosition) < request.payload.targetIndex){
-                request.payload.targetIndex--;
-            }
-            request.payload.targetSelector.spath = request.payload.targetSelector.sid + ':' + request.payload.targetIndex;
-            if(request.payload.targetSelector.id.split('-')[0] === request.payload.targetSelector.sid.split('-')[0]){
-                request.payload.targetSelector.id = request.payload.targetSelector.sid;
-            }
-            let forwardedRequest  = {
-                'payload': {
-                    'selector': request.payload.targetSelector,
-                    'nodeSpec': nodeSpec,
-                    'isMove' : true
-                },
-                'auth': request.auth,
-                'log': request.log.bind(request),
-            };
-            //append the node (revised or not) in the new position
-            self.createDeckTreeNode(forwardedRequest,
-                (inserted) => {
-                    if (inserted.isBoom) return reply(inserted);
-
-                    if(inserted.hasOwnProperty('changeset') && removed.hasOwnProperty('changeset')){
-                        inserted_changeset = inserted.changeset;
-                        inserted.inserted_changeset = inserted_changeset;
-                        inserted.removed_changeset = removed_changeset;
-                    }
-                    else if(removed.hasOwnProperty('changeset')){
-                        inserted.changeset = removed_changeset;
-                    }
-                    if(inserted.hasOwnProperty('changeset')){
-                        inserted_changeset = inserted.changeset;
-                        inserted.changeset = inserted_changeset;
-                    }
-                    reply(inserted);
-                });
-        });
-
     },
 
     getDeckTreeNodeVariants: function(request, reply) {
@@ -2130,7 +1370,7 @@ let self = module.exports = {
     },
 
     getDeckMedia: function(request, reply){
-        deckDB.getMedia(request.params.id, request.query.mediaType).then( (deckMedia) => {
+        treeDB.getMedia(request.params.id, request.query.mediaType).then( (deckMedia) => {
             if(!deckMedia){
                 reply(boom.notFound());
             }
@@ -2197,7 +1437,7 @@ let self = module.exports = {
     regenThumbnails: function(request, reply) {
         let deckId = request.params.id;
 
-        deckDB.getFlatSlides(deckId).then((deckTree) => {
+        treeDB.getFlatSlides(deckId).then((deckTree) => {
             if (!deckTree) return reply(boom.notFound());
 
             async.concatSeries(deckTree.children, (slide, done) => {
