+++ application/database/deckDatabase.js
@@ -1283,42 +1283,6 @@ let self = module.exports = {
         });
     },
 
-    // DEPRECATED
-    updateUsage: function(deck, new_revision_id, root_deck){
-        let idArray = deck.split('-');
-        let rootDeckArray = root_deck.split('-');
-        return helper.connectToDatabase()
-        .then((db) => db.collection('decks'))
-        .then((col) => {
-            return col.findOne({_id: parseInt(idArray[0])})
-            .then((existingDeck) => {
-                //first remove usage of deck from old revision
-                if(root_deck){
-                    let usageArray = existingDeck.revisions[parseInt(idArray[1])-1].usage;
-                    for(let i = 0; i < usageArray.length; i++){
-                        if(usageArray[i].id === parseInt(rootDeckArray[0]) && usageArray[i].revision === parseInt(rootDeckArray[1])){
-                            usageArray.splice(i,1);
-                            break;
-                        }
-                    }
-                    //then update usage array of new/reverted revision
-                    let contains = false;
-                    for(let j = 0; j < existingDeck.revisions[parseInt(new_revision_id)-1].usage.length; j++){
-                        if(existingDeck.revisions[parseInt(new_revision_id)-1].usage[j].id === parseInt(rootDeckArray[0]) && existingDeck.revisions[parseInt(new_revision_id)-1].usage[j].revision === parseInt(rootDeckArray[1])){
-                            contains = true;
-                            break;
-                        }
-                    }
-                    if(!contains)
-                        existingDeck.revisions[parseInt(new_revision_id)-1].usage.push({'id': parseInt(rootDeckArray[0]), 'revision': parseInt(rootDeckArray[1])});
-                }
-                existingDeck.active = new_revision_id;
-                col.save(existingDeck);
-                return existingDeck;
-            });
-        });
-    },
-
     // inserts or updates a variant (language dependent only) id and revision in position index under given deck
     setContentVariant: async function(parentDeckId, index, newVariant, userId) {
         // latest revision always
@@ -1357,92 +1321,6 @@ let self = module.exports = {
         );
 
         return result.value;
-    },
-
-    //recursive function that gets the decktree of a given deck and all of its sub-decks, can be used with onlyDecks to ignore slides
-    // DEPRECATED
-    getDeckTreeFromDB: function(deck_id, onlyDecks){
-        let deckTree;
-        let revision_id = -1;
-        let decktreesplit = String(deck_id).split('-');
-        if(decktreesplit.length > 1){
-            deck_id = decktreesplit[0];
-            revision_id = decktreesplit[1]-1;
-        }
-        return helper.connectToDatabase()
-        .then((db) => db.collection('decks'))
-        .then((col) => {
-            return col.findOne({_id: parseInt(deck_id)})
-            .then((deck) => {
-                if (!deck) return;
-
-                if(revision_id === -1){
-                    revision_id = deck.active-1;
-                }
-                // detect wrong revision id
-                if (!deck.revisions[revision_id]) return;
-
-                let [latestRevision] = deck.revisions.slice(-1);
-
-                deckTree = {
-                    id: deck_id+'-'+(revision_id+1),
-                    revisionId: (revision_id + 1),
-                    latestRevisionId: latestRevision.id,
-                    title: striptags(deck.revisions[revision_id].title),
-                    theme: deck.revisions[revision_id].theme,
-                    allowMarkdown: deck.revisions[revision_id].allowMarkdown,
-                    type: 'deck',
-                    children: [],
-                };
-
-                return new Promise((resolve, reject) => {
-                    async.eachSeries(deck.revisions[revision_id].contentItems, (citem, callback) => {
-                        if(citem.kind === 'slide'){
-                            if(!onlyDecks){
-                                helper.connectToDatabase()
-                                .then((db) => db.collection('slides'))
-                                .then((col) => {
-                                    return col.findOne({_id: parseInt(citem.ref.id)})
-                                    .then((slide) => {
-                                        let slide_revision = citem.ref.revision-1;
-                                        deckTree.children.push({
-                                            title: striptags(slide.revisions[slide_revision].title),
-                                            id: slide._id+'-'+slide.revisions[slide_revision].id,
-                                            type: 'slide',
-                                            theme: deck.revisions[revision_id].theme,
-                                            allowMarkdown: deck.revisions[revision_id].allowMarkdown,
-                                        });
-                                        callback();
-                                    });
-                                }).catch(callback);
-                            }
-                            else{
-                                callback();
-                            }
-                        }
-                        else{
-                            col.findOne({_id: parseInt(citem.ref.id)})
-                            .then((innerDeck) => {
-                                return self.getDeckTreeFromDB(innerDeck._id+'-'+citem.ref.revision, onlyDecks)
-                                .then((res) => {
-                                    deckTree.children.push(res);
-                                    callback();
-                                });
-                            }).catch(callback);
-                        }
-                    }, (err) => {
-                        if (err) {
-                            reject(err);
-                        } else {
-                            resolve(deckTree);
-                        }
-                    });
-
-                });
-
-
-            });
-        });
     },
 
     getSubdeckIds: function(deckId) {
@@ -1497,89 +1375,6 @@ let self = module.exports = {
         });
     },
 
-    //returns a flattened structure of a deck's slides, and optionally its sub-decks
-    // DEPRECATED
-    getFlatSlides: function(deckId, deckTree, returnDecks){
-
-        return self.getRevision(deckId)
-        .then((deckRevision) => {
-
-            // return nothing if not found
-            if (!deckRevision) return;
-
-            if(!deckTree){
-                // info of root deck
-                deckTree = {
-                    title: deckRevision.title,
-                    id: deckId,
-                    type: 'deck',
-                    user: String(deckRevision.user),
-                    theme: deckRevision.theme,
-                    allowMarkdown: deckRevision.allowMarkdown,
-                    children: []
-                };
-            }
-
-            // include subdecks in result
-            if(returnDecks){
-                deckTree.children.push({
-                    title: deckRevision.title,
-                    user: String(deckRevision.user),
-                    id: deckId,
-                    type: 'deck'
-                });
-            }
-
-            return new Promise( (resolve, reject) => {
-                async.eachSeries(deckRevision.contentItems, (citem, callback) => {
-
-                    if(citem.kind === 'slide'){
-                        helper.connectToDatabase()
-                        .then((db) => db.collection('slides'))
-                        .then((col) => {
-                            col.findOne({_id: parseInt(citem.ref.id)})
-                            .then((slide) => {
-                                let slideRevision =  slide.revisions.find((rev) => (rev.id === citem.ref.revision));
-                                deckTree.children.push({
-                                    title:slideRevision.title,
-                                    content: slideRevision.content,
-                                    markdown: slideRevision.markdown,
-                                    speakernotes: slideRevision.speakernotes,
-                                    user: String(slideRevision.user),
-                                    id: slide._id+'-'+slideRevision.id,
-                                    theme: deckRevision.theme,
-                                    allowMarkdown: deckRevision.allowMarkdown,
-                                    type: 'slide'
-                                });
-                                callback();
-                            }).catch( (err) => {
-                                callback(err);
-                            });
-                        }).catch( (err) => {
-                            callback(err);
-                        });
-                    }
-                    else if (citem.kind === 'deck'){
-                        // call recursively for subdecks
-                        self.getFlatSlides(`${citem.ref.id}-${citem.ref.revision}`, deckTree, returnDecks)
-                        .then(() => {
-                            callback();
-                        }).catch( (err) => {
-                            callback(err);
-                        });
-                    }
-                }, (err) => {
-                    if(err){
-                        reject(err);
-                    }
-                    else{
-                        resolve(deckTree);
-                    }
-                });
-            });
-        });
-    },
-
     // returns a flattened structure of a deck's sub-decks
     getFlatDecks: async function(deckId, variantFilter, deckTree) {
         let deck = await self.getDeck(deckId, variantFilter);
@@ -1612,118 +1407,6 @@ let self = module.exports = {
         }
 
         return deckTree;
-    },
-
-    //returns a flattened structure of a deck's sub-decks
-    // DEPRECATED
-    getFlatDecksFromDB: function(deck_id, deckTree){
-
-        let revision_id = -1;
-        let decktreesplit = deck_id.split('-');
-        if(decktreesplit.length > 1){
-            deck_id = decktreesplit[0];
-            revision_id = decktreesplit[1]-1;
-        }
-        return helper.connectToDatabase()
-        .then((db) => db.collection('decks'))
-        .then((col) => {
-            return col.findOne({_id: parseInt(deck_id)})
-            .then((deck) => {
-                if (!deck) return;
-
-                if(revision_id === -1){
-                    revision_id = deck.active-1;
-                }
-                if(!deckTree){
-                    deckTree = { title: deck.revisions[revision_id].title, id: deck_id+'-'+(revision_id+1), type: 'deck', user: String(deck.revisions[revision_id].user), children: []};
-                }
-
-                return new Promise((resolve, reject) => {
-                    async.eachSeries(deck.revisions[revision_id].contentItems, (citem, callback) => {
-
-                        if(citem.kind === 'slide'){
-                            callback();
-                        }
-                        else{
-                            col.findOne({_id: parseInt(citem.ref.id)})
-                            .then((innerDeck) => {
-                                let deck_revision = citem.ref.revision-1;
-                                deckTree.children.push({title: innerDeck.revisions[deck_revision].title, user: String(innerDeck.revisions[deck_revision].user), id: innerDeck._id+'-'+innerDeck.revisions[deck_revision].id, type: 'deck'});
-
-                                return self.getFlatDecksFromDB(innerDeck._id+'-'+citem.ref.revision, deckTree)
-                                .then(() => {
-                                    callback();
-                                });
-                            }).catch(callback);
-                        }
-                    }, (err) => {
-                        if (err) {
-                            reject(err);
-                        } else {
-                            resolve(deckTree);
-                        }
-
-                    });
-
-                });
-            });
-        });
-    },
-
-    // returns an implicit list of editors of a given deck
-    // DEPRECATED
-    getDeckContributors(deck_id, contributorsList){
-        deck_id = String(deck_id);
-        if (!contributorsList) contributorsList = [];
-
-        let revision_id = -1;
-        let decktreesplit = deck_id.split('-');
-        if(decktreesplit.length > 1){
-            deck_id = decktreesplit[0];
-            revision_id = decktreesplit[1]-1;
-        }
-
-        return helper.connectToDatabase()
-        .then((db) => db.collection('decks'))
-        .then((col) => {
-            return col.findOne({_id: parseInt(deck_id)})
-            .then((deck) => {
-                if(revision_id === -1){
-                    revision_id = deck.active-1;
-                }
-
-                // take all revisions older than revision_id (including revision_id)
-                let deckRevisions = deck.revisions.slice(0, revision_id + 1);
-
-                // add all deck revision owners up to and including this revision
-                deckRevisions.forEach((rev) => {
-                    if (rev.user > 0) {
-                        pushIfNotExist(contributorsList, rev.user);
-                    }
-                });
-                pushIfNotExist(contributorsList, deck.user);
-
-                // figure out the subdecks by id and revision
-                let contentItems = deckRevisions.map((rev) => rev.contentItems);
-                contentItems = _.flatten(contentItems).filter((citem) => citem.kind === 'deck');
-                contentItems = _.uniqBy(contentItems, (citem) => `${citem.ref.id}-${citem.ref.revision}`);
-
-                return new Promise((resolve, reject) => {
-                    async.eachSeries(contentItems, (citem, callback) => {
-                        col.findOne({_id: parseInt(citem.ref.id)})
-                        .then((innerDeck) => self.getDeckContributors(innerDeck._id+'-'+citem.ref.revision, contributorsList))
-                        .then(() => callback())
-                        .catch(callback);
-                    }, (err) => {
-                        if (err) {
-                            reject(err);
-                        } else {
-                            resolve(contributorsList);
-                        }
-                    });
-                });
-            });
-        });
     },
 
     // return the set of users and groups with write access to the deck
@@ -1811,312 +1494,6 @@ let self = module.exports = {
             });
 
         });
-
-    },
-
-    // we guard the fork deck revision method against abuse, by checking for change logs of one
-    // DEPRECATED
-    forkDeckRevision(deck_id, user, forAttach) {
-        let deck = util.parseIdentifier(deck_id);
-        return self.get(deck.id).then((existingDeck) => {
-            let [latestRevision] = existingDeck.revisions.slice(-1);
-            if (deck.revision && latestRevision.id !== deck.revision) {
-                // we want to fork a read-only revision, all's well
-                return self._forkDeckRevision(deck_id, user, forAttach);
-            } else {
-                // make the deck id canonical just in case
-                deck.revision = latestRevision.id;
-            }
-
-            // before we fork it, let's check if it's a fresh revision
-            return self.getChangesCounts(deck.id).then((counts) => {
-                if (counts[deck.revision] === 1) {
-                    // we want to fork a fresh revision, let's fork the one before it
-                    console.log(`forking ${deck.revision -1} instead of ${deck.revision} for deck ${deck.id}`);
-                    return self._forkDeckRevision(util.toIdentifier({ id: deck.id, revision: deck.revision - 1 }), user, forAttach);
-                } else {
-                    // unknown revision, old deck without changelog, or a revision with changes, just fork it!
-                    return self._forkDeckRevision(deck_id, user, forAttach);
-                }
-            });
-        });
-    },
-
-    // forks a given deck revision by copying all of its sub-decks into new decks
-    // forAttach is true when forking is done during deck attach process
-    // DEPRECATED
-    _forkDeckRevision(deck_id, user, forAttach) {
-
-        return self.getFlatDecks(deck_id)
-        .then((res) => {
-            //we have a flat sub-deck structure
-            let flatDeckArray = [];
-            flatDeckArray.push(res.id); //push root deck into array
-            for(let i = 0; i < res.children.length; i++){
-                flatDeckArray.push(res.children[i].id); //push next sub-deck into array
-            }
-            //init maps for new ids
-            let id_map = {}, id_noRev_map = {}, slide_id_map = {};
-            //reverse in order to iterate from bottom to top
-            flatDeckArray.reverse();
-            //feed the array for serial processing
-
-            let new_decks = [];
-            return new Promise((resolve, reject) => {
-                //first we generate all the new ids for the copied decks, and hold them in a map for future reference
-                async.eachSeries(flatDeckArray, (next_deck, callback) => {
-                    return helper.connectToDatabase()
-                    .then((db) => helper.getNextIncrementationValueForCollection(db, 'decks'))
-                    .then((newId) => {
-                        id_map[next_deck] = newId+'-'+1;
-                        id_noRev_map[next_deck.split('-')[0]] = newId;
-                        callback();
-                    }).catch(callback);
-                }, (err) => {
-                    if (err) {
-                        return reject(err);
-                    }
-
-                    //iterate the flat decktree and copy each deck, referring to the new ids in its content items and usage
-                    async.eachSeries(flatDeckArray, (next_deck, callback) => {
-                        let {id: nextId, revision: nextRevision} = util.parseIdentifier(next_deck);
-
-                        helper.connectToDatabase() //db connection have to be accessed again in order to work with more than one collection
-                        .then((db2) => db2.collection('decks'))
-                        .then((col) => {
-                            return col.findOne({_id: nextId})
-                            .then((found) => {
-                                let ind = _.findIndex(found.revisions, { id: nextRevision });
-                                let contributorsArray = found.contributors;
-                                //contributorsArray.push({'user': parseInt(user), 'count': 1});
-                                let existingUserContributorIndex = findWithAttr(contributorsArray, 'user', parseInt(user));
-                                if(existingUserContributorIndex > -1)
-                                    contributorsArray[existingUserContributorIndex].count++;
-                                else{
-                                    contributorsArray.push({'user': parseInt(user), 'count': 1});
-                                }
-
-                                let copiedDeck = {
-                                    _id: id_noRev_map[found._id],
-                                    origin: {
-                                        id: found._id,
-                                        revision: found.revisions[ind].id,
-                                        title: found.revisions[ind].title,
-                                        user: found.user,
-                                    },
-                                    description: found.description,
-                                    language: found.revisions[ind].language,
-                                    license: found.license,
-                                    user: parseInt(user),
-                                    translated_from: found.translated_from,
-                                    contributors: contributorsArray,
-                                    active: 1,
-                                    // TODO revisit how we maintain this attribute
-                                    translations: found.translations || [],
-                                    // forked decks are created as hidden like they were new ones
-                                    hidden: true,
-                                };
-                                if (found.slideDimensions) {
-                                    copiedDeck.slideDimensions = found.slideDimensions;
-                                }
-
-                                let now = new Date();
-                                let timestamp = now.toISOString();
-                                copiedDeck.timestamp = timestamp;
-                                copiedDeck.lastUpdate = timestamp;
-                                if(found.hasOwnProperty('datasource')){
-                                    copiedDeck.datasource = found.datasource;
-                                }
-                                else{
-                                    copiedDeck.datasource = null;
-                                }
-                                //copiedDeck.parent = next_deck.split('-')[0]+'-'+next_deck.split('-')[1];
-                                copiedDeck.revisions = [found.revisions[ind]];
-                                copiedDeck.revisions[0].id = 1;
-                                // own the revision as well!
-                                copiedDeck.revisions[0].user = copiedDeck.user;
-
-                                // renew creation date for fresh revision
-                                copiedDeck.revisions[0].timestamp = timestamp;
-                                copiedDeck.revisions[0].lastUpdate = timestamp;
-
-                                // this points to the same deck, needs to be removed in forked decks
-                                delete copiedDeck.revisions[0].originRevision;
-
-                                // isFeatured needs to be removed in forked decks
-                                delete copiedDeck.revisions[0].isFeatured;
-
-                                let contentItemsMap = {};
-                                let contentItemsToCopy = []; // TODO for now disable copying slides as well // copiedDeck.revisions[0].contentItems;
-                                let copiedDeckId = util.toIdentifier({id: copiedDeck._id, revision: 1});
-
-                                // TODO remove this circular reference
-                                let slideDB = require('./slideDatabase');
-                                async.eachSeries(contentItemsToCopy, (nextSlide, doneSlideCopy) => {
-                                    if (nextSlide.kind !== 'slide') return doneSlideCopy();
-
-                                    // let's copy the slide
-                                    return helper.getCollection('slides').then((slides) => {
-                                        return slides.findOne({_id: nextSlide.ref.id}).then((slide) => {
-                                            let oldSlideId = slide._id;
-                                            let sourceRevision = _.find(slide.revisions, { id: nextSlide.ref.revision });
-                                            let newSlide = Object.assign({}, slide, sourceRevision, { id: slide._id, revision: sourceRevision.id });
-                                            return slideDB.copy(newSlide, copiedDeckId, parseInt(user)).then((inserted) => {
-                                                let newSlideId = inserted._id;
-                                                contentItemsMap[oldSlideId] = newSlideId;
-                                                slide_id_map[oldSlideId] = newSlideId;
-
-                                                // create the thumbnail
-                                                let copiedSlideId = `${newSlideId}-1`;
-                                                fileService.createThumbnail(inserted.revisions[0].content, copiedSlideId, copiedDeck.revisions[0].theme).catch((err) => {
-                                                    console.warn(`could not create thumbnail for translation ${copiedSlideId}, error was: ${err.message}`);
-                                                });
-
-                                                //console.log('contentItemsMap', contentItemsMap);
-                                                //console.log('copiedDeck', copiedDeck);
-                                                for(let i = 0; i < contentItemsToCopy.length; i++){
-                                                    if(contentItemsToCopy[i].ref.id === oldSlideId){
-                                                        contentItemsToCopy[i].ref.id = newSlideId;
-                                                        contentItemsToCopy[i].ref.revision = 1;
-                                                    }
-                                                }
-
-                                                doneSlideCopy();
-                                            });
-
-                                        });
-                                    }).catch(doneSlideCopy); // catch ANY error
-
-                                }, (err) => {
-                                    if (err) {
-                                        return callback(err);
-                                    }
-
-                                    // these should be run with or without a translation operation
-                                    for(let i = 0; i < copiedDeck.revisions[0].contentItems.length; i++){
-                                        for(let j in id_map){
-                                            if(id_map.hasOwnProperty(j) && copiedDeck.revisions[0].contentItems[i].ref.id === parseInt(j.split('-')[0])){
-                                                copiedDeck.revisions[0].contentItems[i].ref.id = parseInt(id_map[j].split('-')[0]);
-                                                copiedDeck.revisions[0].contentItems[i].ref.revision = parseInt(id_map[j].split('-')[1]);
-                                            }
-                                        }
-                                    }
-                                    for(let i = 0; i < copiedDeck.revisions[0].usage.length; i++){
-                                        for(let j in id_map){
-                                            if(id_map.hasOwnProperty(j) && copiedDeck.revisions[0].usage[i].id === parseInt(j.split('-')[0])){
-                                                copiedDeck.revisions[0].usage[i].id = parseInt(id_map[j].split('-')[0]);
-                                                copiedDeck.revisions[0].usage[i].revision = parseInt(id_map[j].split('-')[1]);
-                                            }
-                                        }
-                                    }
-                                    for(let i = 0; i < copiedDeck.revisions[0].contentItems.length; i++){
-                                        let nextSlide = copiedDeck.revisions[0].contentItems[i];
-                                        //console.log('nextSlide', nextSlide);
-                                        if(nextSlide.kind === 'slide'){
-                                            let root_deck_path = [copiedDeck._id, '1'];
-                                            //console.log('outside root_deck_path', root_deck_path);
-                                            //console.log('contentItemsMap', contentItemsMap);
-                                            self.addToUsage(nextSlide, root_deck_path);
-                                        }
-                                        else{
-                                            continue;
-                                        }
-                                    }
-
-                                    new_decks.push(copiedDeck);
-                                    return col.insertOne(copiedDeck).then(() => {
-                                        callback();
-                                    }).catch(callback);
-
-                                });
-
-                            });
-
-                        }).catch(callback);
-
-                    }, (err) => {
-                        if (err) {
-                            reject(err);
-                        } else {
-                            if (!forAttach) {
-                                // if not attaching, we need to track stuff here
-                                let rootDeckId = id_map[res.id];
-                                self._trackDecksForked(rootDeckId, id_map, user, 'fork');
-                            }
-
-                            resolve({'root_deck': id_map[res.id], 'id_map': id_map});
-                        }
-                    });
-                });
-            });
-
-        }).then((forkResult) => {
-            // after forking the deck and if the revision we forked is the latest,
-            // we create a new revision for the original deck;
-            // this way the fork points to a read-only revision
-
-            let deck = util.parseIdentifier(deck_id);
-            return self.get(deck.id).then((existingDeck) => {
-                let [latestRevision] = existingDeck.revisions.slice(-1);
-                if (deck.revision && latestRevision.id !== deck.revision) {
-                    // we forked a read-only revision, nothing to do here
-                    return forkResult;
-                } else {
-                    // make the deck id canonical just in case
-                    deck.revision = latestRevision.id;
-                }
-
-                // this is an automatic revision, the user should be 'system'
-                // deck autorevision is created with same deck as root
-                return self.createDeckRevision(deck.id, -1, deck.id).then((updatedDeck) => {
-                    // we need to update all parents of the deck to keep them updated
-                    // with the latest revision we have just created now
-                    return self.getUsage(util.toIdentifier(deck)).then((usage) => {
-                        // if a deck has no roots, itself is the root
-                        console.log(`updating deck revision used for ${deck.id} in ${usage.length} parent decks`);
-
-                        usage.reduce((p, parentDeck) => {
-                            return p.then(() => {
-                                // citem, revertedRevId, root_deck, ckind, user, top_root_deck, parentOperations
-                                let parentDeckId = util.toIdentifier(parentDeck);
-                                return self.updateContentItem(updatedDeck, '', parentDeckId, 'deck', -1, parentDeckId);
-                            });
-                        }, Promise.resolve());
-                    }).then(() => {
-                        // return the same result
-                        return forkResult;
-                    });
-
-                });
-
-            });
-
-        });
-
-    },
-
-    // TODO make this actually private after code in handler.js has been moved here
-    _trackDecksForked(rootDeckId, forkIdsMap, userId, forkType='fork') {
-        // we reverse the array to track the root first, then the children in order
-        let newDeckIds = Object.keys(forkIdsMap).map((key) => forkIdsMap[key]).reverse();
-
-        let parentOperations = [];
-        // taken from https://stackoverflow.com/questions/30823653/is-node-js-native-promise-all-processing-in-parallel-or-sequentially/#30823708
-        // this starts with a promise that resolves to empty array,
-        // then takes each new deck id and applies the tracking and returns a new promise that resolves
-        // to the tracking results, that are picked up by the next iteration, etc...
-        return newDeckIds.reduce((p, newDeckId) => {
-            return p.then((deckChanges) => {
-                // if errored somewhere return nothing, chain will just end without doing the rest
-                if (!deckChanges) return;
-
-                // parent operations is only the ops for the forking of the first deck (the root of the fork tree)
-                // the first time this runs, deckChanges is empty!
-                if (_.isEmpty(parentOperations)) parentOperations.push(...deckChanges);
-                // we track everything as rooted to the deck_id
-                return ChangeLog.trackDeckForked(newDeckId, userId, rootDeckId, parentOperations, forkType);
-            });
-        }, Promise.resolve([]));
 
     },
 
@@ -2724,24 +2101,6 @@ let self = module.exports = {
             .then((updated) => updated.value);
         });
 
-    },
-
-    // fetches specified media-type files that are present inside the deck
-    getMedia: function(deckId, mediaType){
-        return self.getFlatSlides(deckId, undefined, false).then( (flatSlides) => {
-            if(!flatSlides) return;
-
-            // get media uris per slide as arrays
-            let media = flatSlides.children.map( (slide) => {
-                return util.findMedia(slide.content, mediaType);
-            });
-
-            // flatten arrays of media uris
-            let flatMedia = [].concat.apply([], media);
-
-            // return unique media uris
-            return [...new Set(flatMedia)];
-        });
     },
 
     // fetches change log records for the deck or subdecks thereof
